Rewrite h5p modals to use separate modals in separate files
